Removes ITextModel dependency from objectStream.ts
